move meta out to its own file
